Enable runtime filters on multi CN
