+++ pkg/sql/compile/compile.go
@@ -1530,8 +1530,7 @@ func (c *Compile) compileTableScanWithNode(n *plan.Node, node engine.Node, filte
 	s.Proc = process.NewWithAnalyze(c.proc, c.ctx, 0, c.anal.Nodes())
 
 	// Register runtime filters
-	// XXX currently we only enable runtime filter on single CN
-	if len(c.cnList) == 1 && len(n.RuntimeFilterProbeList) > 0 {
+	if len(n.RuntimeFilterProbeList) > 0 {
 		receivers := make([]*colexec.RuntimeFilterChan, len(n.RuntimeFilterProbeList))
 		if c.runtimeFilterReceiverMap == nil {
 			c.runtimeFilterReceiverMap = make(map[int32]chan *pipeline.RuntimeFilter)
