+++ pkg/sql/compile/scopeRemoteRun.go
@@ -226,7 +226,12 @@ func receiveMessageFromCnServer(c *Compile, sender *messageSenderOnClient, nextA
 		}
 		sequence++
 
+		//if dataBuffer == nil {
+		//	dataBuffer = m.Data
+		//} else {
 		dataBuffer = append(dataBuffer, m.Data...)
+		//}
+
 		if m.WaitingNextToMerge() {
 			continue
 		}
@@ -444,13 +449,13 @@ func generatePipeline(s *Scope, ctx *scopeContext, ctxId int32) (*pipeline.Pipel
 			}
 			p.DataSource.Block = string(data)
 		}
-		//if len(s.DataSource.RuntimeFilterReceivers) > 0 {
-		//	rfSpecs := make([]*plan.RuntimeFilterSpec, len(s.DataSource.RuntimeFilterReceivers))
-		//	for i, recv := range s.DataSource.RuntimeFilterReceivers {
-		//		rfSpecs[i] = recv.Spec
-		//	}
-		//	p.DataSource.RuntimeFilterList = rfSpecs
-		//}
+		if len(s.DataSource.RuntimeFilterReceivers) > 0 {
+			rfSpecs := make([]*plan.RuntimeFilterSpec, len(s.DataSource.RuntimeFilterReceivers))
+			for i, recv := range s.DataSource.RuntimeFilterReceivers {
+				rfSpecs[i] = recv.Spec
+			}
+			p.DataSource.RuntimeFilterList = rfSpecs
+		}
 	}
 	// PreScope
 	p.Children = make([]*pipeline.Pipeline, len(s.PreScopes))
@@ -560,21 +565,21 @@ func generateScope(proc *process.Process, p *pipeline.Pipeline, ctx *scopeContex
 			bat.Cnt = 1
 			s.DataSource.Bat = bat
 		}
-		//if len(dsc.RuntimeFilterList) > 0 {
-		//	rfReceivers := make([]*colexec.RuntimeFilterChan, len(dsc.RuntimeFilterList))
-		//	if ctx.runtimeFilterReceiverMap == nil {
-		//		ctx.runtimeFilterReceiverMap = make(map[int32]chan *pipeline.RuntimeFilter)
-		//	}
-		//	for i, rfSpec := range dsc.RuntimeFilterList {
-		//		ch := make(chan *pipeline.RuntimeFilter, 1)
-		//		rfReceivers[i] = &colexec.RuntimeFilterChan{
-		//			Spec: rfSpec,
-		//			Chan: ch,
-		//		}
-		//		ctx.runtimeFilterReceiverMap[rfSpec.Tag] = ch
-		//	}
-		//	s.DataSource.RuntimeFilterReceivers = rfReceivers
-		//}
+		if len(dsc.RuntimeFilterList) > 0 {
+			rfReceivers := make([]*colexec.RuntimeFilterChan, len(dsc.RuntimeFilterList))
+			if ctx.runtimeFilterReceiverMap == nil {
+				ctx.runtimeFilterReceiverMap = make(map[int32]chan *pipeline.RuntimeFilter)
+			}
+			for i, rfSpec := range dsc.RuntimeFilterList {
+				ch := make(chan *pipeline.RuntimeFilter, 1)
+				rfReceivers[i] = &colexec.RuntimeFilterChan{
+					Spec: rfSpec,
+					Chan: ch,
+				}
+				ctx.runtimeFilterReceiverMap[rfSpec.Tag] = ch
+			}
+			s.DataSource.RuntimeFilterReceivers = rfReceivers
+		}
 	}
 	if p.Node != nil {
 		s.NodeInfo.Id = p.Node.Id
@@ -953,13 +958,13 @@ func convertToPipelineInstruction(opr *vm.Instruction, ctx *scopeContext, ctxId 
 			Types:    convertToPlanTypes(t.Typs),
 			Conds:    t.Conditions,
 		}
-		//if len(t.RuntimeFilterSenders) > 0 {
-		//	rfSpecs := make([]*plan.RuntimeFilterSpec, len(t.RuntimeFilterSenders))
-		//	for i, sender := range t.RuntimeFilterSenders {
-		//		rfSpecs[i] = sender.Spec
-		//	}
-		//	in.HashBuild.RuntimeFilterList = rfSpecs
-		//}
+		if len(t.RuntimeFilterSenders) > 0 {
+			rfSpecs := make([]*plan.RuntimeFilterSpec, len(t.RuntimeFilterSenders))
+			for i, sender := range t.RuntimeFilterSenders {
+				rfSpecs[i] = sender.Spec
+			}
+			in.HashBuild.RuntimeFilterList = rfSpecs
+		}
 	case *external.Argument:
 		name2ColIndexSlice := make([]*pipeline.ExternalName2ColIndex, len(t.Es.Name2ColIndex))
 		i := 0
@@ -1321,26 +1326,26 @@ func convertToVmInstruction(opr *pipeline.Instruction, ctx *scopeContext, eng en
 		}
 	case vm.HashBuild:
 		t := opr.GetHashBuild()
-		//var rfSenders []*colexec.RuntimeFilterChan
-		//if t.RuntimeFilterList != nil {
-		//	rfSenders = make([]*colexec.RuntimeFilterChan, 0, len(t.RuntimeFilterList))
-		//	for _, rfSpec := range t.RuntimeFilterList {
-		//		if ch, ok := ctx.runtimeFilterReceiverMap[rfSpec.Tag]; ok {
-		//			rfSenders = append(rfSenders, &colexec.RuntimeFilterChan{
-		//				Spec: rfSpec,
-		//				Chan: ch,
-		//			})
-		//		}
-		//	}
-		//}
+		var rfSenders []*colexec.RuntimeFilterChan
+		if t.RuntimeFilterList != nil {
+			rfSenders = make([]*colexec.RuntimeFilterChan, 0, len(t.RuntimeFilterList))
+			for _, rfSpec := range t.RuntimeFilterList {
+				if ch, ok := ctx.runtimeFilterReceiverMap[rfSpec.Tag]; ok {
+					rfSenders = append(rfSenders, &colexec.RuntimeFilterChan{
+						Spec: rfSpec,
+						Chan: ch,
+					})
+				}
+			}
+		}
 		v.Arg = &hashbuild.Argument{
-			Ibucket:     t.Ibucket,
-			Nbucket:     t.Nbucket,
-			NeedHashMap: t.NeedHash,
-			NeedExpr:    t.NeedExpr,
-			Typs:        convertToTypes(t.Types),
-			Conditions:  t.Conds,
-			//RuntimeFilterSenders: rfSenders,
+			Ibucket:              t.Ibucket,
+			Nbucket:              t.Nbucket,
+			NeedHashMap:          t.NeedHash,
+			NeedExpr:             t.NeedExpr,
+			Typs:                 convertToTypes(t.Types),
+			Conditions:           t.Conds,
+			RuntimeFilterSenders: rfSenders,
 		}
 	case vm.External:
 		t := opr.GetExternalScan()
