+++ pkg/vm/engine/types.go
@@ -25,7 +25,6 @@ import (
 	"github.com/matrixorigin/matrixone/pkg/container/batch"
 	"github.com/matrixorigin/matrixone/pkg/container/types"
 	"github.com/matrixorigin/matrixone/pkg/container/vector"
-	"github.com/matrixorigin/matrixone/pkg/pb/pipeline"
 	"github.com/matrixorigin/matrixone/pkg/pb/plan"
 	"github.com/matrixorigin/matrixone/pkg/pb/timestamp"
 	"github.com/matrixorigin/matrixone/pkg/txn/client"
@@ -517,8 +516,6 @@ type Relation interface {
 	Statistics
 
 	Ranges(context.Context, []*plan.Expr) ([][]byte, error)
-
-	ApplyRuntimeFilters(context.Context, [][]byte, []*plan.Expr, []*pipeline.RuntimeFilter) ([][]byte, error)
 
 	TableDefs(context.Context) ([]TableDef, error)
